requirements.txt
c # Dependencias principales
groq>=0.0.2
python-dotenv>=1.0.0
loguru>=0.7.2
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.2

# Desarrollo
pytest>=8.0.0
pytest-cov>=4.1.0
black>=24.0.0
isort>=5.13.0
flake8>=6.1.0
mypy>=1.8.0

# Documentación
mkdocs>=1.5.3
mkdocs-material>=9.5.0
